refactor: 피드백 반영 - 템플릿메서드 패턴 적용
