+++ src/main/kotlin/game/blackjack/domain/Dealer.kt
@@ -1,16 +1,17 @@
 package game.blackjack.domain
 
-class Dealer : Player("딜러", 0) {
+class Dealer : Participant("딜러") {
     private val deck = Deck()
 
     fun drawCard(): Card = deck.draw()
+
     fun drawCard(count: Int): List<Card> {
         return (1..count).map { drawCard() }
     }
 
     override fun receiveUntilHit(
         action: (name: String) -> Boolean,
-        showPlayerCard: (player: Player) -> Unit,
+        showPlayerCard: (participant: Participant) -> Unit,
         drawCard: () -> Card
     ) {
         while (hand.score() <= CAN_RECEIVE_SCORE) {
