+++ src/main/kotlin/game/blackjack/domain/Player.kt
@@ -1,28 +1,16 @@
 package game.blackjack.domain
 
-open class Player(val name: String, val money: Int) {
-
-    val hand: Hand = Hand()
+class Player(name: String, val money: Int) : Participant(name) {
 
     constructor(name: String) : this(name, 0)
 
     fun determine(response: Boolean) = hand.setStatus(if (response) Status.HIT else Status.STAY)
 
-    fun score(): Score = hand.score()
-
-    fun isBust(): Boolean = hand.isBust()
-
-    fun isBlackJack(): Boolean = hand.isBlackJack()
-
-    fun init(cards: List<Card>) = hand.init(cards)
-
-    fun receive(card: Card) = hand.receive(card)
-
     fun canReceive(): Boolean = hand.isHit()
 
-    open fun receiveUntilHit(
+    override fun receiveUntilHit(
         action: (name: String) -> Boolean,
-        showPlayerCard: (player: Player) -> Unit,
+        showPlayerCard: (participant: Participant) -> Unit,
         drawCard: () -> Card
     ) {
         while (canReceive()) {
