+++ src/main/kotlin/game/blackjack/domain/Players.kt
@@ -11,7 +11,7 @@ class Players(
         return this
     }
 
-    fun distribute(getAction: (name: String) -> Boolean, showPlayerCard: (player: Player) -> Unit): Players {
+    fun distribute(getAction: (name: String) -> Boolean, showPlayerCard: (participant: Participant) -> Unit): Players {
         (players + dealer).forEach {
             it.receiveUntilHit(getAction, showPlayerCard) { dealer.drawCard() }
         }
@@ -20,7 +20,7 @@ class Players(
 
     fun forEach(action: (player: Player) -> Unit) = players.forEach { action(it) }
 
-    fun forEachWithDealer(action: (player: Player) -> Unit) =
+    fun forEachWithDealer(action: (participant: Participant) -> Unit) =
         listOf(dealer, *(players.toTypedArray())).forEach { action(it) }
 
     fun getResult(): Map<String, Int> {
